Add files via upload
This is a data science project on an advertising dataset. The dataset displays how different advertising mediums affect the sales of a particular company (https://www.statlearning.com/resources-second-edition). The approach I used to identify the relationships between advertising and sales is linear regression (and multi-linear regression). There are a number of comments above the code that explain my approach.